Move all address commands' parser tests classes to address sub-package
