+++ src/test/java/seedu/jarvis/logic/parser/address/AddAddressCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.jarvis.logic.parser;
+package seedu.jarvis.logic.parser.address;
 
 import static seedu.jarvis.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.jarvis.logic.commands.CommandTestUtil.ADDRESS_DESC_AMY;
@@ -32,7 +32,6 @@ import static seedu.jarvis.testutil.TypicalPersons.BOB;
 import org.junit.jupiter.api.Test;
 
 import seedu.jarvis.logic.commands.address.AddAddressCommand;
-import seedu.jarvis.logic.parser.address.AddAddressCommandParser;
 import seedu.jarvis.model.person.Address;
 import seedu.jarvis.model.person.Email;
 import seedu.jarvis.model.person.Name;
