+++ src/test/java/seedu/jarvis/logic/parser/address/DeleteAddressCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.jarvis.logic.parser;
+package seedu.jarvis.logic.parser.address;
 
 import static seedu.jarvis.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.jarvis.logic.parser.CommandParserTestUtil.assertParseFailure;
@@ -8,7 +8,6 @@ import static seedu.jarvis.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import org.junit.jupiter.api.Test;
 
 import seedu.jarvis.logic.commands.address.DeleteAddressCommand;
-import seedu.jarvis.logic.parser.address.DeleteAddressCommandParser;
 
 /**
  * As we are only doing white-box testing, our test cases do not cover path variations
